==> 2_Feature Enginnering.ipynb <==
### This scripts takes about 3 minutes seconds to execute

import sqlalchemy as db
import pandas as pd
import numpy as np
import datetime as dt
from math import isnan
import sklearn.preprocessing as preprocessing
pd.options.display.max_columns = None

# Load runners and races from Smartform database

from sqlalchemy import text

engine = db.create_engine('mysql+pymysql://smartform:smartform@localhost/smartform')
conn = engine.connect()

query = text("""
SELECT * from new_my_runners;
""")

runners = pd.read_sql_query(query, conn, parse_dates=['meeting_date'],
                       dtype={'stall_number':np.int64, 'trainer_id':np.int64, 'sire_id': np.int64, 'dam_id': np.int64, 'finpos': np.int64, 'win': np.int64})

query = text("""
SELECT * from new_my_races;
""")

races = pd.read_sql_query(query, conn, parse_dates=['meeting_date'])

# Sort racing data in order of ascending meeting dates, e.g., to ensure daysLTO's are calculated sensibly
runners.sort_values(by=['runner_id', 'meeting_date', 'race_id'], inplace=True)

### blinkers, visor, cheekpieces, or tonguetie (each 1=yes if they were wearing these; 0=no)

runners['blinkers'] = runners.apply(lambda row: 0 if isnan(row['tack_blinkers']) else 1, axis=1)
runners['visor'] = runners.apply(lambda row: 0 if isnan(row['tack_visor']) else 1, axis=1)
runners['cheekpieces'] = runners.apply(lambda row: 0 if isnan(row['tack_cheek_piece']) else 1, axis=1)
runners['tonguetie'] = runners.apply(lambda row: 0 if isnan(row['tack_tongue_strap']) else 1, axis=1)

del runners['tack_blinkers']
del runners['tack_visor']
del runners['tack_cheek_piece']
del runners['tack_tongue_strap']

### entire - male horse that has not been castrated (1=yes, 0=no)
### gelding - male horse that has been castrated (1=yes, 0=no)
### note that a horse that is neither a gelding or an entire was female

runners['entire'] = runners.gender.apply(lambda g: 0 if g in ['G', 'F', 'M'] else 1)
runners['gelding'] = runners.gender.apply(lambda g: 0 if g in ['F', 'M', 'C', 'H'] else 1)

### daysLTO - days since Last Time Out (capped at 1 year; replace missing values with 1 year)
### position1 - finishing position in the previous race (1, 2, 3 or 4, 0 = anywhere else)
### position2 - finishing position two races ago (1, 2, 3 or 4, 0 = anywhere else)
### position3 - finishing position three races ago (1, 2, 3 or 4, 0 = anywhere else)

def calc_basic_form(runners_split):
    n = len(runners_split)
    daysLTO = np.full(n, 365, dtype=np.int64) # default to 1 year
    position1 = np.full(n, 0, dtype=np.int64) # default to 0
    position2 = np.full(n, 0, dtype=np.int64) # default to 0
    position3 = np.full(n, 0, dtype=np.int64) # default to 0
    
    for i in range(n):
        if i > 0:
            daysLTO[i] = (runners_split.meeting_date.iloc[i] - runners_split.meeting_date.iloc[i-1]).days
            daysLTO[i] = min(365, daysLTO[i]) # cap at 1 year
            position1[i] = min(4, runners_split.finpos.iloc[i-1])

        if i > 1:
            position2[i] = min(4, runners_split.finpos.iloc[i-2])

        if i > 2:
            position3[i] = min(4, runners_split.finpos.iloc[i-3])

    return pd.DataFrame(zip(runners_split.race_id, daysLTO, position1, position2, position3), columns=['race_id', 'daysLTO', 'position1', 'position2', 'position3'])

runners_temp = runners.groupby('runner_id').apply(calc_basic_form)
runners_temp.reset_index(inplace=True)
del runners_temp['level_1'] # proxy for race_id
runners = pd.merge(runners, runners_temp, how='left', on=['runner_id', 'race_id'], validate='1:1')

# Visualize features

%matplotlib widget
import matplotlib.pyplot as plt
from matplotlib.ticker import FuncFormatter

# convenience function
def make_chart(feature, title, axis=None, missing=None, kind="barh", thousands=True):
    if missing:
        runners_temp = pd.DataFrame(np.unique(['Missing' if isnan(feature) else 'Not Missing' for feature in runners[feature]], return_counts=True)).transpose()
    else:
        runners_temp = pd.DataFrame(np.unique(runners[feature], return_counts=True)).transpose()
    runners_temp.columns = [feature, 'count']
    runners_temp.set_index(feature, drop=True, inplace=True)

    if kind == "barh":
        thingy = runners_temp.plot
    elif kind == "hist":
        thingy = runners[feature].plot
    else:
        raise ValueError('A very specific bad thing happened.') 
    
    if axis == None:
        ax = thingy(kind=kind)
    else:
        ax = thingy(ax=axis, kind=kind)
    ax.set_title(title)
    
    if kind == "barh":
        ax.set_xlabel('Frequency')
        ax.get_xaxis().set_major_formatter(
            FuncFormatter(lambda x, p: format(int(x), ',')))
    elif kind == "hist":
        ax.set_xlabel(feature)
        if thousands:
            ax.get_xaxis().set_major_formatter(
                FuncFormatter(lambda x, p: format(int(x), ',')))
        ax.get_yaxis().set_major_formatter(
            FuncFormatter(lambda y, p: format(int(y), ',')))
    else:
        raise ValueError('A very specific bad thing happened.') 

make_chart('age', 'Age of horse (years)')

make_chart('finpos', '(Amended) finishing position of horse')

make_chart('win', 'Indicator of whether horse won (1) or not (0)')

fig, axes = plt.subplots(1, 3, figsize=(14,6))
make_chart('sp', 'Starting prices w/ outliers', kind="hist", axis=axes[0])
axes[1].boxplot(runners.sp)
axes[1].set_ylabel('sp')
axes[1].set_title('Starting Prices w/ outliers')
axes[2].boxplot(runners.sp, showfliers=False)
axes[2].set_ylabel('sp')
axes[2].set_title('Starting prices w/o outliers')
fig.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.25, hspace=0.5)

fig, axes = plt.subplots(2, 2, figsize=(14,12))
make_chart('blinkers', 'Was the horse wearing blinkers (1=yes, 0=no)?', axis=axes[0][0])
make_chart('visor', 'Was the horse wearing visor (1=yes, 0=no)?', axis=axes[0][1])
make_chart('cheekpieces', 'Was the horse wearing cheekpieces (1=yes, 0=no)?', axis=axes[1][0])
make_chart('tonguetie', 'Was the horse wearing tonguetie (1=yes, 0=no)?', axis=axes[1][1])
fig.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.25, hspace=0.25)

fig, axes = plt.subplots(2, 2, figsize=(14,12))
make_chart('entire', 'Was the horse entire (1=yes, 0=no)?', axis=axes[0][0])
make_chart('gelding', 'Was the horse gelding (1=yes, 0=no)?', axis=axes[0][1])
make_chart('gender', 'Horse gender (not used in probability model)', axis=axes[1][0])
axes[1][1].axis('off')
axes[1][1].text(0, 0.8, "Letter representing sex of horse (G)elding, (F)illie, (M)are, (C)olt, (H)orse", wrap=True)
fig.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.25, hspace=0.25)

fig, axes = plt.subplots(2, 2, figsize=(14,12))
make_chart('position1', 'Position in prior race', axes[0][0])
make_chart('position2', 'Position two races prior', axes[0][1])
make_chart('position3', 'Position three races prior', axes[1][0])
make_chart('daysLTO', 'Days since last time out (daysLTO)', kind="hist", axis=axes[1][1])
fig.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.25, hspace=0.25)

# note, most daysLTO's are non-zero; the few that are, c. 50, result from those horses running two races on the same date
np.nanpercentile(runners.daysLTO, q=(0, 1,2, 3, 4, 5))

# note, win percentages are as of latest meeting date; therefore, if horse has single outing he will have zero win percentage
fig, axes = plt.subplots(1, 2, figsize=(14,6))
make_chart('win_perc', 'Win percentages', kind="hist", axis=axes[0], thousands=False)
make_chart('win_perc', 'Win percentages', missing=True, axis=axes[1])
fig.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.25, hspace=0.25)

# note, sire and dam strike strikes are as of latest meeting date; therefore, if horse has single outing he will have missing sire and dam strikes rates;
# he will also have missing strikes rates if his sire or dam aren't part of the All Weather dataset.
fig, axes = plt.subplots(2, 2, figsize=(14,12))
make_chart('sire_sr', 'Sire strikes rates', kind="hist", axis=axes[0][0], thousands=False)
make_chart('sire_sr', 'Sire strike rates', missing=True, axis=axes[0][1])
make_chart('dam_sr', 'Dam strikes rates', kind="hist", axis=axes[1][0], thousands=False)
make_chart('dam_sr', 'Dam strike rates', missing=True, axis=axes[1][1])
fig.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.25, hspace=0.25)

# note, trainer strike rates are as of latest meeting date; therefore, if horse has single outing he will have missing trainer strike rates;
# he will also have missing strikes rates if his trainer isn't part of the All Weather dataset
fig, axes = plt.subplots(1, 2, figsize=(14,6))
make_chart('trainer_sr', 'Trainer strike rates', kind="hist", axis=axes[0], thousands=False)
make_chart('trainer_sr', 'Trainer strikes rates', missing=True, axis=axes[1])
fig.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.25, hspace=0.25)

## Additional variables for model MG1 and other versions of this model

fig, axes = plt.subplots(2, 3, figsize=(14,12))
make_chart('sr_30d', "Runner's 30-day strike rate", axes[0][0], kind="hist", thousands=False)
make_chart('sr_lifetime_class_5', "Runner's lifetime class-5 strike rate", kind="hist", axis=axes[0][1], thousands=False)
make_chart('sr_lifetime_class_4', "Runner's lifetime class-4 strike rate", kind="hist", axis=axes[0][2], thousands=False)
make_chart('sr_lifetime_class_3', "Runner's lifetime class-3 strike rate", kind="hist", axis=axes[1][0], thousands=False)
make_chart('sr_lifetime_class_2', "Runner's lifetime class-2 strike rate", kind="hist", axis=axes[1][1], thousands=False)
make_chart('sr_lifetime_class_1', "Runner's lifetime class-1 strike rate", kind="hist", axis=axes[1][2], thousands=False)
fig.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.25, hspace=0.25)

fig, axes = plt.subplots(2, 3, figsize=(14,12))
make_chart('jockey_sr_30d', "Jockey's 30-day strike rate", kind="hist", axis=axes[0][0], thousands=False)
make_chart('jockey_sr_KEM', "Jockey's strike rate at Kempton", axes[0][1], kind="hist", thousands=False)
make_chart('jockey_sr_LIN', "Jockey's strike rate at Linfield", kind="hist", axis=axes[0][2], thousands=False)
make_chart('jockey_sr_SOU', "Jockey's strike rate at Southwell", kind="hist", axis=axes[1][0], thousands=False)
make_chart('jockey_sr_WOL', "Jockey's strike rate at Wolverhampton", kind="hist", axis=axes[1][1], thousands=False)
fig.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.25, hspace=0.25)

fig, axes = plt.subplots(2, 3, figsize=(14,12))
make_chart('trainer_sr_30d', "Trainer's 30-day strike rate", kind="hist", axis=axes[0][0], thousands=False)
make_chart('trainer_sr_KEM', "Trainer's strike rate at Kempton", axes[0][1], kind="hist", thousands=False)
make_chart('trainer_sr_LIN', "Trainer's strike rate at Linfield", kind="hist", axis=axes[0][2], thousands=False)
make_chart('trainer_sr_SOU', "Trainer's strike rate at Southwell", kind="hist", axis=axes[1][0], thousands=False)
make_chart('trainer_sr_WOL', "Trainer's strike rate at Wolverhampton", kind="hist", axis=axes[1][1], thousands=False)
fig.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.25, hspace=0.25)

fig, axes = plt.subplots(2, 3, figsize=(14,12))
make_chart('pos_prior1_KEM_le1mi', "Runners's pos. prior, <=1mi, Kem.", kind="hist", axis=axes[0][0], thousands=False)
make_chart('pos_prior2_KEM_le1mi', "Runners's pos. prior 2, <=1mi, Kem.", kind="hist", axis=axes[0][1], thousands=False)
make_chart('pos_prior3_KEM_le1mi', "Runners's pos. prior 3, <=1mi, Kem.", kind="hist", axis=axes[0][2], thousands=False)
make_chart('pos_prior1_LIN_le1mi', "Runners's pos. prior, <=1mi, Lin.", kind="hist", axis=axes[1][0], thousands=False)
make_chart('pos_prior2_LIN_le1mi', "Runners's pos. prior 2, <=1mi, Lin.", kind="hist", axis=axes[1][1], thousands=False)
make_chart('pos_prior3_LIN_le1mi', "Runners's pos. prior 3, <=1mi, Lin.", kind="hist", axis=axes[1][2], thousands=False)
fig.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.25, hspace=0.25)

fig, axes = plt.subplots(2, 3, figsize=(14,12))
make_chart('pos_prior1_SOU_le1mi', "Runners's pos. prior, <=1mi, Sou.", kind="hist", axis=axes[0][0], thousands=False)
make_chart('pos_prior2_SOU_le1mi', "Runners's pos. prior 2, <=1mi, Sou.", kind="hist", axis=axes[0][1], thousands=False)
make_chart('pos_prior3_SOU_le1mi', "Runners's pos. prior 3, <=1mi, Sou.", kind="hist", axis=axes[0][2], thousands=False)
make_chart('pos_prior1_WOL_le1mi', "Runners's pos. prior, <=1mi, Wol.", kind="hist", axis=axes[1][0], thousands=False)
make_chart('pos_prior2_WOL_le1mi', "Runners's pos. prior 2, <=1mi, Wol.", kind="hist", axis=axes[1][1], thousands=False)
make_chart('pos_prior3_WOL_le1mi', "Runners's pos. prior 3, <=1mi, Wol.", kind="hist", axis=axes[1][2], thousands=False)
fig.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.25, hspace=0.25)

fig, axes = plt.subplots(2, 3, figsize=(14,12))
make_chart('pos_prior1_KEM_gt1mi', "Runners's pos. prior, >1mi, Kem.", kind="hist", axis=axes[0][0], thousands=False)
make_chart('pos_prior2_KEM_gt1mi', "Runners's pos. prior 2, >1mi, Kem.", kind="hist", axis=axes[0][1], thousands=False)
make_chart('pos_prior3_KEM_gt1mi', "Runners's pos. prior 3, >1mi, Kem.", kind="hist", axis=axes[0][2], thousands=False)
make_chart('pos_prior1_LIN_gt1mi', "Runners's pos. prior, >1mi, Lin.", kind="hist", axis=axes[1][0], thousands=False)
make_chart('pos_prior2_LIN_gt1mi', "Runners's pos. prior 2, >1mi, Lin.", kind="hist", axis=axes[1][1], thousands=False)
make_chart('pos_prior3_LIN_gt1mi', "Runners's pos. prior 3, >1mi, Lin.", kind="hist", axis=axes[1][2], thousands=False)
fig.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.25, hspace=0.25)

fig, axes = plt.subplots(2, 3, figsize=(14,12))
make_chart('pos_prior1_SOU_gt1mi', "Runners's pos. prior, >1mi, Sou.", kind="hist", axis=axes[0][0], thousands=False)
make_chart('pos_prior2_SOU_gt1mi', "Runners's pos. prior 2, >1mi, Sou.", kind="hist", axis=axes[0][1], thousands=False)
make_chart('pos_prior3_SOU_gt1mi', "Runners's pos. prior 3, >1mi, Sou.", kind="hist", axis=axes[0][2], thousands=False)
make_chart('pos_prior1_WOL_gt1mi', "Runners's pos. prior, >1mi, Wol.", kind="hist", axis=axes[1][0], thousands=False)
make_chart('pos_prior2_WOL_gt1mi', "Runners's pos. prior 2, >1mi, Wol.", kind="hist", axis=axes[1][1], thousands=False)
make_chart('pos_prior3_WOL_gt1mi', "Runners's pos. prior 3, >1mi, Lwol.", kind="hist", axis=axes[1][2], thousands=False)
fig.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.25, hspace=0.25)

### Default missing sire, dam, and trainer strikes rates to zero, and put cap on these rates

runners['sire_sr'] = runners['sire_sr'].apply(lambda sr: 0 if isnan(sr) else sr)
runners['dam_sr'] = runners['dam_sr'].apply(lambda sr: 0 if isnan(sr) else sr)
runners['trainer_sr'] = runners['trainer_sr'].apply(lambda sr: 0 if isnan(sr) else sr)

runners['sire_sr'] = runners['sire_sr'].apply(lambda sr: min(0.4, sr)) # 40% cap
runners['dam_sr'] = runners['dam_sr'].apply(lambda sr: min(0.4, sr)) # 40% cap
runners['trainer_sr'] = runners['trainer_sr'].apply(lambda sr: min(0.4, sr)) # 40% cap

fig, axes = plt.subplots(2, 2, figsize=(14,12))
make_chart('sire_sr', 'Sire strikes rates', kind="hist", axis=axes[0][0], thousands=False)
make_chart('sire_sr', 'Sire strike rates', missing=True, axis=axes[0][1])
make_chart('dam_sr', 'Dam strikes rates', kind="hist", axis=axes[1][0], thousands=False)
make_chart('dam_sr', 'Dam strike rates', missing=True, axis=axes[1][1])
fig.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.25, hspace=0.25)

fig, axes = plt.subplots(1, 2, figsize=(14,6))
make_chart('trainer_sr', 'Trainer strike rates', kind="hist", axis=axes[0], thousands=False)
make_chart('trainer_sr', 'Trainer strikes rates', missing=True, axis=axes[1])
fig.subplots_adjust(left=None, bottom=None, right=None, top=None, wspace=0.25, hspace=0.25)

print("All Weather racecourses in the UK at {}".format(", ".join(np.unique(races.course))))
print("{:,} horses taking part in {:,} races".format(len(np.unique(runners.runner_id)), len(np.unique(runners.race_id))))
races.groupby('course')['race_id'].count()

# Create dummy variables

runners_temp = pd.get_dummies(runners[["position1", "position2", "position3"]], columns=["position1", "position2", "position3"], drop_first=True)
runners = pd.concat([runners, runners_temp], axis=1)
del runners_temp

races_temp = pd.get_dummies(races[["course", "class"]], columns=["course", "class"], drop_first=False)
races = pd.concat([races, races_temp], axis=1)
del races_temp

# Add other indicator variables

races['gt1mi'] = np.where(races['distance_yards'] > 1760, 1, 0)

# Divide races into training, validation, and testing datasets

cutoff_1 = dt.datetime(2012, 10, 1)
cutoff_2 = dt.datetime(2014, 2, 1)

runners_train = runners[runners['meeting_date'] < cutoff_1]
runners_train.shape

races_train = races[races['meeting_date'] < cutoff_1]
races_train.shape

runners_validate = runners[np.logical_and(runners['meeting_date'] >= cutoff_1, runners['meeting_date'] < cutoff_2)]
runners_validate.shape

races_validate = races[np.logical_and(races['meeting_date'] >= cutoff_1, races['meeting_date'] < cutoff_2)]
races_validate.shape

runners_test = runners[runners['meeting_date'] >= cutoff_2]
runners_test.shape

races_test = races[races['meeting_date'] >= cutoff_2]
races_test.shape

print('{:.0%} of races will be used for training the model'.format(races_train.shape[0]/races.shape[0]))
print('{:.0%} of races will be used for validating the model'.format(races_validate.shape[0]/races.shape[0]))
print('{:.0%} of races will be used for testing the model'.format(races_test.shape[0]/races.shape[0]))

# Save data

runners_train.to_csv('data\\runners_train.csv', index=False)
runners_validate.to_csv('data\\runners_validate.csv', index=False)
runners_test.to_csv('data\\runners_test.csv', index=False)
runners.to_csv('data\\runners.csv', index=False)

races_train.to_csv('data\\races_train.csv', index=False)
races_validate.to_csv('data\\races_validate.csv', index=False)
races_test.to_csv('data\\races_test.csv', index=False)
races.to_csv('data\\races.csv', index=False)

# another convenience function
def make_chart2(values, index, ax, title1, title2):
    ax.plot(pd.pivot_table(runners_train, values=values, index=index, aggfunc='mean'))
    ax.set_xlabel(title1)
    ax.set_ylabel('(mean) win proportion')
    ax.set_title(title2)

# Examine relationships between win percentages and features

### Win percentage versus age (training set)

plt.rcdefaults()
fig = plt.figure(tight_layout=True, figsize=(14,6))

make_chart2("win_perc", ['age'], fig.add_subplot(111), 'age (years)', 'Win proportion vs. age (training set)')

### Win percentage versus sire strike rate, dam strike rate, trainer strike rate, and days since last time out (training set)

# convert to whole percentages
runners_train = runners_train.assign(sire_sr_rounded = np.around(runners_train['sire_sr']/10, decimals=2)*10*100)
runners_train = runners_train.assign(dam_sr_rounded = np.around(runners_train['dam_sr']/10, decimals=2)*10*100)
runners_train = runners_train.assign(trainer_sr_rounded = np.around(runners_train['trainer_sr']/10, decimals=2)*10*100)
runners_train = runners_train.assign(daysLTO_rounded = np.around(runners['daysLTO']/30, decimals=0)*30)
runners_train = runners_train.assign(patrick=0)

fig = plt.figure(tight_layout=True, figsize=(14,12))

make_chart2("win_perc", ['sire_sr_rounded'], fig.add_subplot(221), 'strike rate (%)', 'Win perc vs. sire SR to nearest 10% (training set)')
make_chart2("win_perc", ['dam_sr_rounded'], fig.add_subplot(222), 'strike rate (%)', 'Win perc vs. dam SR to nearest 10% (training set)')
make_chart2("win_perc", ['trainer_sr_rounded'], fig.add_subplot(223), 'strike rate (%)', 'Win perc vs. trainer SR to nearest 10% (training set)')
make_chart2("win_perc", ['daysLTO_rounded'], fig.add_subplot(224), 'strike rate (%)', 'Win perc vs. daysLTO to nearest 30 days (training set)')

# clean up
del runners_train['sire_sr_rounded']
del runners_train['dam_sr_rounded']
del runners_train['trainer_sr_rounded']
del runners_train['daysLTO_rounded']
del runners_train['patrick']

### Win percentage vs. position in the horse's previous races (training set)

fig = plt.figure(tight_layout=True)

ax = fig.add_subplot(111)
ax.plot(pd.pivot_table(runners_train, values="win_perc", index=['position1'], aggfunc='mean'), "--o", label="position1")
ax.plot(pd.pivot_table(runners_train, values="win_perc", index=['position2'], aggfunc='mean'), "--o", label="position2")
ax.plot(pd.pivot_table(runners_train, values="win_perc", index=['position3'], aggfunc='mean'), "--o", label="position3")
ax.set_xlabel('position')
ax.set_ylabel('(mean) win percentage')
ax.set_title('Win percentage vs. position (training set)')
ax.legend()

### Win percentage versus entire, geldin, blinkers, visor, cheekpieces or tongue-tie (training set)

entire = pd.pivot_table(runners_train, values="win_perc", columns=["entire"], aggfunc='mean')
gelding = pd.pivot_table(runners_train, values="win_perc", columns=["gelding"], aggfunc='mean')
blinkers = pd.pivot_table(runners_train, values="win_perc", columns=["blinkers"], aggfunc='mean')
visor = pd.pivot_table(runners_train, values="win_perc", columns=["visor"], aggfunc='mean')
cheekpieces = pd.pivot_table(runners_train, values="win_perc", columns=["cheekpieces"], aggfunc='mean')
tonguetie = pd.pivot_table(runners_train, values="win_perc", columns=["tonguetie"], aggfunc='mean')

pd.DataFrame([[float(entire[1]), float(gelding[1]), float(blinkers[1]), float(visor[1]), float(cheekpieces[1]), float(tonguetie[1])], [float(entire[0]), float(gelding[0]), float(blinkers[0]), float(visor[0]), float(cheekpieces[0]), float(tonguetie[0])]],
             index=["yes", "no"], columns=["entire", "gelding", "blinkers", "visor", "cheekpieces", "tonguetie"])

## Additional variables for model MG1 and other versions of this model

### Win percentage versus runner's, jockey's, and trainer's 30-day strike rates (training set)

# convert to whole percentages
runners_train = runners_train.assign(sr_30d_rounded = np.around(runners_train['sr_30d']/10, decimals=2)*10*100)
runners_train = runners_train.assign(jockey_sr_30d_rounded = np.around(runners_train['jockey_sr_30d']/10, decimals=2)*10*100)
runners_train = runners_train.assign(trainer_sr_30d_rounded = np.around(runners_train['trainer_sr_30d']/10, decimals=2)*10*100)

fig = plt.figure(tight_layout=True, figsize=(14,12))

make_chart2("win_perc", ['sr_30d_rounded'], fig.add_subplot(221), 'strike rate (%)', 'Win perc vs. runner 30d SR to nearest 10% (training set)')
make_chart2("win_perc", ['jockey_sr_30d_rounded'], fig.add_subplot(222), 'strike rate (%)', 'Win perc vs. jockey 30d SR to nearest 10% (training set)')
make_chart2("win_perc", ['trainer_sr_30d_rounded'], fig.add_subplot(223), 'strike rate (%)', 'Win perc vs. trainer 30d SR to nearest 10% (training set)')

# clean up
del runners_train['sr_30d_rounded']
del runners_train['jockey_sr_30d_rounded']
del runners_train['trainer_sr_30d_rounded']


### Win percentage versus runner's lifetime strike rates at different classes (training set)

# convert to whole percentages
runners_train = runners_train.assign(sr_lifetime_class_5_rounded = np.around(runners_train['sr_lifetime_class_5']/10, decimals=2)*10*100)
runners_train = runners_train.assign(sr_lifetime_class_4_rounded = np.around(runners_train['sr_lifetime_class_4']/10, decimals=2)*10*100)
runners_train = runners_train.assign(sr_lifetime_class_3_rounded = np.around(runners_train['sr_lifetime_class_3']/10, decimals=2)*10*100)
runners_train = runners_train.assign(sr_lifetime_class_2_rounded = np.around(runners_train['sr_lifetime_class_2']/10, decimals=2)*10*100)
runners_train = runners_train.assign(sr_lifetime_class_1_rounded = np.around(runners_train['sr_lifetime_class_1']/10, decimals=2)*10*100)

fig = plt.figure(tight_layout=True, figsize=(14,12))

make_chart2("win_perc", ['sr_lifetime_class_5_rounded'], fig.add_subplot(231), 'strike rate (%)', 'Win perc vs. class 5 SR to nearest 10% (training set)')
make_chart2("win_perc", ['sr_lifetime_class_4_rounded'], fig.add_subplot(232), 'strike rate (%)', 'Win perc vs. class 4 SR to nearest 10% (training set)')
make_chart2("win_perc", ['sr_lifetime_class_3_rounded'], fig.add_subplot(233), 'strike rate (%)', 'Win perc vs. class 3 SR to nearest 10% (training set)')
make_chart2("win_perc", ['sr_lifetime_class_2_rounded'], fig.add_subplot(234), 'strike rate (%)', 'Win perc vs. class 2 SR to nearest 10% (training set)')
make_chart2("win_perc", ['sr_lifetime_class_1_rounded'], fig.add_subplot(235), 'strike rate (%)', 'Win perc vs. class 1 SR to nearest 10% (training set)')

# clean up
del runners_train['sr_lifetime_class_5_rounded']
del runners_train['sr_lifetime_class_4_rounded']
del runners_train['sr_lifetime_class_3_rounded']
del runners_train['sr_lifetime_class_2_rounded']
del runners_train['sr_lifetime_class_1_rounded']

### Win percentage versus jockey's lifetime strike rates at different tracks (training set)

# convert to whole percentages
runners_train = runners_train.assign(jockey_sr_KEM_rounded = np.around(runners_train['jockey_sr_KEM']/10, decimals=2)*10*100)
runners_train = runners_train.assign(jockey_sr_LIN_rounded = np.around(runners_train['jockey_sr_LIN']/10, decimals=2)*10*100)
runners_train = runners_train.assign(jockey_sr_SOU_rounded = np.around(runners_train['jockey_sr_SOU']/10, decimals=2)*10*100)
runners_train = runners_train.assign(jockey_sr_WOL_rounded = np.around(runners_train['jockey_sr_WOL']/10, decimals=2)*10*100)

fig = plt.figure(tight_layout=True, figsize=(14,12))

make_chart2("win_perc_KEM", ['jockey_sr_KEM_rounded'], fig.add_subplot(221), 'strike rate (%)', 'Win perc vs. jockey KEM SR to nearest 10% (training set)')
make_chart2("win_perc_LIN", ['jockey_sr_LIN_rounded'], fig.add_subplot(222), 'strike rate (%)', 'Win perc vs. jockey LIN SR to nearest 10% (training set)')
make_chart2("win_perc_SOU", ['jockey_sr_SOU_rounded'], fig.add_subplot(223), 'strike rate (%)', 'Win perc vs. jockey SOU SR to nearest 10% (training set)')
make_chart2("win_perc_WOL", ['jockey_sr_WOL_rounded'], fig.add_subplot(224), 'strike rate (%)', 'Win perc vs. jockey WOL SR to nearest 10% (training set)')

# clean up
del runners_train['jockey_sr_KEM_rounded']
del runners_train['jockey_sr_LIN_rounded']
del runners_train['jockey_sr_SOU_rounded']
del runners_train['jockey_sr_WOL_rounded']

### Win percentage versus trainer's lifetime strike rates at different tracks (training set)

# convert to whole percentages
runners_train = runners_train.assign(trainer_sr_KEM_rounded = np.around(runners_train['trainer_sr_KEM']/10, decimals=2)*10*100)
runners_train = runners_train.assign(trainer_sr_LIN_rounded = np.around(runners_train['trainer_sr_LIN']/10, decimals=2)*10*100)
runners_train = runners_train.assign(trainer_sr_SOU_rounded = np.around(runners_train['trainer_sr_SOU']/10, decimals=2)*10*100)
runners_train = runners_train.assign(trainer_sr_WOL_rounded = np.around(runners_train['trainer_sr_WOL']/10, decimals=2)*10*100)

fig = plt.figure(tight_layout=True, figsize=(14,12))

make_chart2("win_perc_KEM", ['trainer_sr_KEM_rounded'], fig.add_subplot(221), 'strike rate (%)', 'Win perc vs. trainer KEM SR to nearest 10% (training set)')
make_chart2("win_perc_LIN", ['trainer_sr_LIN_rounded'], fig.add_subplot(222), 'strike rate (%)', 'Win perc vs. trainer LIN SR to nearest 10% (training set)')
make_chart2("win_perc_SOU", ['trainer_sr_SOU_rounded'], fig.add_subplot(223), 'strike rate (%)', 'Win perc vs. trainer SOU SR to nearest 10% (training set)')
make_chart2("win_perc_WOL", ['trainer_sr_WOL_rounded'], fig.add_subplot(224), 'strike rate (%)', 'Win perc vs. trainer WOL SR to nearest 10% (training set)')

# clean up
del runners_train['trainer_sr_KEM_rounded']
del runners_train['trainer_sr_LIN_rounded']
del runners_train['trainer_sr_SOU_rounded']
del runners_train['trainer_sr_WOL_rounded']

### Win percentage vs. position in the horse's previous races at or under 1 mile at Kempton (training set)

fig = plt.figure(tight_layout=True)

ax = fig.add_subplot(111)
ax.plot(pd.pivot_table(runners_train, values="win_perc_KEM", index=['pos_prior1_KEM_le1mi'], aggfunc='mean'), "--o", label="prior1")
ax.plot(pd.pivot_table(runners_train, values="win_perc_KEM", index=['pos_prior2_KEM_le1mi'], aggfunc='mean'), "--o", label="prior2")
ax.plot(pd.pivot_table(runners_train, values="win_perc_KEM", index=['pos_prior3_KEM_le1mi'], aggfunc='mean'), "--o", label="prior3")
ax.set_xlabel('position')
ax.set_ylabel('(mean) win percentage')
ax.set_title('Win percentage vs. position KEM, <= 1mi (training set)')
ax.legend()

### Win percentage vs. position in the horse's previous races at or under 1 mile at Lingfield (training set)

fig = plt.figure(tight_layout=True)

ax = fig.add_subplot(111)
ax.plot(pd.pivot_table(runners_train, values="win_perc_LIN", index=['pos_prior1_LIN_le1mi'], aggfunc='mean'), "--o", label="prior1")
ax.plot(pd.pivot_table(runners_train, values="win_perc_LIN", index=['pos_prior2_LIN_le1mi'], aggfunc='mean'), "--o", label="prior2")
ax.plot(pd.pivot_table(runners_train, values="win_perc_LIN", index=['pos_prior3_LIN_le1mi'], aggfunc='mean'), "--o", label="prior3")
ax.set_xlabel('position')
ax.set_ylabel('(mean) win percentage')
ax.set_title('Win percentage vs. position LIN, <= 1mi (training set)')
ax.legend()

### Win percentage vs. position in the horse's previous races at or under 1 mile at Southwell (training set)

fig = plt.figure(tight_layout=True)

ax = fig.add_subplot(111)
ax.plot(pd.pivot_table(runners_train, values="win_perc_SOU", index=['pos_prior1_SOU_le1mi'], aggfunc='mean'), "--o", label="prior1")
ax.plot(pd.pivot_table(runners_train, values="win_perc_SOU", index=['pos_prior2_SOU_le1mi'], aggfunc='mean'), "--o", label="prior2")
ax.plot(pd.pivot_table(runners_train, values="win_perc_SOU", index=['pos_prior3_SOU_le1mi'], aggfunc='mean'), "--o", label="prior3")
ax.set_xlabel('position')
ax.set_ylabel('(mean) win percentage')
ax.set_title('Win percentage vs. position SOU, <= 1mi (training set)')
ax.legend()

### Win percentage vs. position in the horse's previous races at or under 1 mile at Wolverhampton (training set)

fig = plt.figure(tight_layout=True)

ax = fig.add_subplot(111)
ax.plot(pd.pivot_table(runners_train, values="win_perc_WOL", index=['pos_prior1_WOL_le1mi'], aggfunc='mean'), "--o", label="prior1")
ax.plot(pd.pivot_table(runners_train, values="win_perc_WOL", index=['pos_prior2_WOL_le1mi'], aggfunc='mean'), "--o", label="prior2")
ax.plot(pd.pivot_table(runners_train, values="win_perc_WOL", index=['pos_prior3_WOL_le1mi'], aggfunc='mean'), "--o", label="prior3")
ax.set_xlabel('position')
ax.set_ylabel('(mean) win percentage')
ax.set_title('Win percentage vs. position WOL, <= 1mi (training set)')
ax.legend()

### Win percentage vs. position in the horse's previous races over 1 mile at Kempton (training set)

fig = plt.figure(tight_layout=True)

ax = fig.add_subplot(111)
ax.plot(pd.pivot_table(runners_train, values="win_perc_KEM", index=['pos_prior1_KEM_gt1mi'], aggfunc='mean'), "--o", label="prior1")
ax.plot(pd.pivot_table(runners_train, values="win_perc_KEM", index=['pos_prior2_KEM_gt1mi'], aggfunc='mean'), "--o", label="prior2")
ax.plot(pd.pivot_table(runners_train, values="win_perc_KEM", index=['pos_prior3_KEM_gt1mi'], aggfunc='mean'), "--o", label="prior3")
ax.set_xlabel('position')
ax.set_ylabel('(mean) win percentage')
ax.set_title('Win percentage vs. position KEM, > 1mi (training set)')
ax.legend()

### Win percentage vs. position in the horse's previous races over 1 mile at Lingfield (training set)

fig = plt.figure(tight_layout=True)

ax = fig.add_subplot(111)
ax.plot(pd.pivot_table(runners_train, values="win_perc_LIN", index=['pos_prior1_LIN_gt1mi'], aggfunc='mean'), "--o", label="prior1")
ax.plot(pd.pivot_table(runners_train, values="win_perc_LIN", index=['pos_prior2_LIN_gt1mi'], aggfunc='mean'), "--o", label="prior2")
ax.plot(pd.pivot_table(runners_train, values="win_perc_LIN", index=['pos_prior3_LIN_gt1mi'], aggfunc='mean'), "--o", label="prior3")
ax.set_xlabel('position')
ax.set_ylabel('(mean) win percentage')
ax.set_title('Win percentage vs. position LIN, > 1mi (training set)')
ax.legend()

### Win percentage vs. position in the horse's previous races over 1 mile at Southwell (training set)

fig = plt.figure(tight_layout=True)

ax = fig.add_subplot(111)
ax.plot(pd.pivot_table(runners_train, values="win_perc_SOU", index=['pos_prior1_SOU_gt1mi'], aggfunc='mean'), "--o", label="prior1")
ax.plot(pd.pivot_table(runners_train, values="win_perc_SOU", index=['pos_prior2_SOU_gt1mi'], aggfunc='mean'), "--o", label="prior2")
ax.plot(pd.pivot_table(runners_train, values="win_perc_SOU", index=['pos_prior3_SOU_gt1mi'], aggfunc='mean'), "--o", label="prior3")
ax.set_xlabel('position')
ax.set_ylabel('(mean) win percentage')
ax.set_title('Win percentage vs. position SOU, > 1mi (training set)')
ax.legend()

### Win percentage vs. position in the horse's previous races over 1 mile at Wolverhampton (training set)

fig = plt.figure(tight_layout=True)

ax = fig.add_subplot(111)
ax.plot(pd.pivot_table(runners_train, values="win_perc_WOL", index=['pos_prior1_WOL_gt1mi'], aggfunc='mean'), "--o", label="prior1")
ax.plot(pd.pivot_table(runners_train, values="win_perc_WOL", index=['pos_prior2_WOL_gt1mi'], aggfunc='mean'), "--o", label="prior2")
ax.plot(pd.pivot_table(runners_train, values="win_perc_WOL", index=['pos_prior3_WOL_gt1mi'], aggfunc='mean'), "--o", label="prior3")
ax.set_xlabel('position')
ax.set_ylabel('(mean) win percentage')
ax.set_title('Win percentage vs. position WOL, > 1mi (training set)')
ax.legend()
